src/domain/forum/application/use-cases/edit-answer.ts
import { Answer } from '../../enterprise/entities/answer'
import { AnswersRepository } from '../repositories/answers-repository'

interface EditAnswerUseCaseDTO {
  authorId: string
  answerId: string
  content: string
}

interface EditAnswerUseCaseResponse {
  answer: Answer
}

class EditAnswerUseCase {
  constructor(private readonly answerRepository: AnswersRepository) {}

  async execute({
    answerId,
    authorId,
    content,
  }: EditAnswerUseCaseDTO): Promise<EditAnswerUseCaseResponse> {
    const answer = await this.answerRepository.findById(answerId)

    if (!answer) {
      throw new Error('Answer not found')
    }

    if (authorId !== answer.authorId.toString()) {
      throw new Error('You are not allowed to edit this answer')
    }

    answer.content = content

    await this.answerRepository.update(answer)

    return {
      answer,
    }
  }
}

export { EditAnswerUseCase }
